backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.0
langchain-huggingface==0.0.1
langchain-groq==0.0.1
faiss-cpu==1.7.4
pytesseract==0.3.10
Pillow==10.1.0
requests==2.31.0
youtube-transcript-api==0.6.2